Hopf debug
Hopf model code in python has been debugged. the initial value has been added in the first column of the array and the loop is now doing 999 iterations. I also added all the K matrix for each gaits. the initial value setting has been changed to match the matlab code. even if both way works. I also corrected the equation for omega because I did a mistake when writing it last time. I also tried to make the plot more readable to enhance clarity ( especially when two neurons are in phase). the only thing missing is how neurons are inhibited/connected together to check that gaits are correct and follow the definition of the gaits. All the gaits gives good output with correct phase and amplitude