debug Stein
added comment to  each gait to explain what is wrong and how they should be coupled ( in phase or out of phase) according to the gait definition and Collins and Richmond 1994. Since all three gaits are wrong it might be that the loop/equations are wrong. I will try to fix it later.